gm-fastapi-streamlit/src/fastapi/requirements.txt
alembic==1.11.1
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
databricks==0.2
databricks-api==0.8.0
databricks-cli==0.17.7
databricks-sql-connector==2.5.2
ecdsa==0.18.0
et-xmlfile==1.1.0
fastapi==0.95.2
fastapi-azure-auth==4.1.1
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
lz4==4.3.2
Mako==1.2.4
MarkupSafe==2.1.3
numpy==1.24.3
oauthlib==3.2.2
openpyxl==3.1.2
pandas==1.5.3
passlib==1.7.4
pyarrow==12.0.0
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.8
PyJWT==2.7.0
pymssql==2.2.7
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2023.3
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.48
starlette==0.27.0
tabulate==0.9.0
thrift==0.16.0
typing_extensions==4.6.2
urllib3==1.26.16
uvicorn==0.22.0
